Move Markdown facade into Facades namespace
