+++ packages/framework/src/Facades/Includes.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Facades;
 
 use Hyde\Framework\Contracts\IncludeFacadeContract;
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Modules\Markdown\Markdown;
+use Hyde\Framework\Facades\Markdown;
 use Illuminate\Support\Facades\Blade;
 
 class Includes implements IncludeFacadeContract
